Breaking: Remove YGConfigGetInstanceCount
Summary:
This was added in https://github.com/facebook/yoga/pull/497 specifically for tests related to memory leaks in the C# bindings to count how often YGConfigFree.

This is the wrong layer for this check, we don't have officially supported C# bindings anymore, and this API is not safe when Yoga runs on multiple threads. This removes it.

Differential Revision: D49131207

fbshipit-source-id: fd31e294655ab7d37a81adf7ea6c3f8d3e87c0d6