Fixup exports
Summary:
D18029030 added `-fvisibility-hidden`, and a corresponding `YOGA_EXPORT` macro for defining shared library visibility.

There was already a `WIN_EXPORT` macro doing the same thing when building a DLL, but in the header, instead of CPP files, and sometimes missing for important parts of the public API?

This reconciles those into a single visiblity macro, that we always place with declaration instead of definition.

Differential Revision: D49132643

fbshipit-source-id: 27dc91380707ff289c1e0ff88cce24b5e84a1e00