Breaking: Remove YGConfigGetInstanceCount (#1370)
Summary:
Pull Request resolved: https://github.com/facebook/yoga/pull/1370

X-link: https://github.com/facebook/react-native/pull/39369

This was added in https://github.com/facebook/yoga/pull/497 specifically for tests related to memory leaks in the C# bindings to count how often YGConfigFree.

This is the wrong layer for this check, we don't have officially supported C# bindings anymore, and this API is not safe when Yoga runs on multiple threads. This removes it, similar to a global node instance count that was also previously removed.

Reviewed By: rshest

Differential Revision: D49131207

fbshipit-source-id: 4e0f15068a10f95597e3421d3e7ad84d840eec94