Use repo root babel config in /javascript (#1509)
Summary:
Pull Request resolved: https://github.com/facebook/yoga/pull/1509

D51966201 adds a new Babel config to the repo root. This makes it so the existing JS package shares that, and shifts around the dependencies a bit, to avoid reliance on our specific hoisting resolution.

Babel's config structure makes it annoying to be able to resolve a babel config across a package boundary unless you declare your own next to where it may be executed, so I kept a babel config where it was before, just referencing the root one.

Docusaurus is stil using its own `docusaurus/core/lib/babel/preset`.

Differential Revision: D52070490

fbshipit-source-id: 0c211b2cbe19ade08867c9980fb922f628e0980e