Add command line options to new node-based gentest
Summary: This uses the minimist library to parse the command line easily for flags and options. I added over the same ones in the ruby script (-f and -s). Unfortunately, there seems to be no way to fully suspend the script on command while waiting for user input like how you can in ruby. So, I am only able to suspend at the very end by not closing the browser immediately. Therefore, I only let this happen if -f is specified.

Differential Revision: https://www.internalfb.com/diff/D51877810?entry_point=27

fbshipit-source-id: 7ae1b2cac8ff7253854f71fae5747bdcb2366ba5