Create node version of gentest script
Summary:
The only instance of ruby in this repository is `gentest.rb` used to generate test cases from html fixtures. This is quite annoying as ruby is not the most popular compared to something like Node and it does not integrate into the rest of our stack. I changed this to use Node.js instead. Instead of `watir` we now use `selenium-webdriver`. `watir` is backed by Selenium so I do not expect anything to change.

Next commits will add command line options, clean up gentest.rb and its references, and change the README

Differential Revision: https://www.internalfb.com/diff/D51874433?entry_point=27

fbshipit-source-id: 586a201cb5741241d2402f9dfd5bc382863e8d89