Remove static-behaves-like-relative errata
Summary:
Since we aim to ship static to all users of yoga (not just XPR), we need to remove the errata that is gating most of the features. For the most part, this is a non breaking change because static is no longer the default as of D52137858. However, there is one tidbit that is changing:

The height and width of the parent is changing from `availableInnerHeight/Width` to the padding box of the node. While it varies, `availableInnerHeight/Width` is usually the content box, although what is important is that it is different than the padding box. What will happen with this change is that absolute nodes with percentage based values with be different if their parent has padding.

The [padding box is the correct size that should be passed into absolute layout](https://developer.mozilla.org/en-US/docs/Web/CSS/Containing_block#identifying_the_containing_block). If we do not want to ship this change (i.e. if too many things break as a result), we could in theory pass in `availableInnerHeight/Width` to `layoutAbsoluteDescendants` and do some funky if's to replicate the behavior. However, I would rather try to make this correct once and for all.

Differential Revision: https://www.internalfb.com/diff/D52706161?entry_point=27

fbshipit-source-id: b3369f40a55edd10c2e38ccec31aafad150cfad7