Ability to capture trees as JSON files (#1565)
Summary:

This adds a `captureTree` method intended to be injected in somewhere where you would like a JSON representation of a yoga tree (like in calculateLayout). Right now it is very simple and just calls on `nodeToString` to serialize a node into JSON, and then saves that into a file in the `benchmark` dir. Some buck file changes needed to be done as well as this is the first `.cpp` file in `benchmark`

Differential Revision: D52972995