Ability to capture trees as JSON files (#1565)
Summary:
Pull Request resolved: https://github.com/facebook/yoga/pull/1565

This adds a `captureTree` method intended to be injected in somewhere where you would like a JSON representation of a yoga tree. Right now it is very simple and just calls on `nodeToString` to serialize a node into JSON, and then saves that into a file in the given path. Some buck file changes needed to be done as well to use this in other files.

Reviewed By: NickGerleman

Differential Revision: D52972995

fbshipit-source-id: f4e09a815edef92ab959cfc76bacccbce225d940