Expose box sizing getters and setters in Yoga
Summary: I would like to write some tests for box sizing that will drive a lot of my development as I implement content box. To do that, I need this publicly exposed. Obviously not that ideal since this currently does not do anything. Maybe we can name the value in such a way that its clear it is in development?

Reviewed By: NickGerleman

Differential Revision: D63135970