Add "SPM Supported" Badge to README
Summary:
CocoaPods is EOL (though we will still support it for a while), and Yoga has SPM support, so let's add a badge!

SPM does not have a central registry (it instead uses our Git branch), so there isn't a badge to show the current vesion, we instead just mention that we support it, which is what I found other projects do.

Differential Revision: D63845470