Add "SPM Supported" Badge to README (#1717)
Summary:
Pull Request resolved: https://github.com/facebook/yoga/pull/1717

CocoaPods is EOL (though we will still support it for a while), and Yoga has SPM support, so let's add a badge!

SPM does not have a central registry (it instead uses our Git branch), so there isn't a badge to show the current vesion, we instead just mention that we support it, which is what I found other projects do.

Reviewed By: nmn

Differential Revision: D63845470

fbshipit-source-id: d033700fa04b8d708890b1c51d0a37fab8251a43